Unify ListenBrainz client
