+++ sidetrack/api/api/v1/listens.py
@@ -13,8 +13,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sidetrack.common.models import Artist, Listen, Track, UserSettings
 
 from ...clients.lastfm import LastfmClient, get_lastfm_client
-from ...clients.listenbrainz import ListenBrainzClient, get_listenbrainz_client
 from ...clients.spotify import SpotifyClient, get_spotify_client
+from sidetrack.services.listenbrainz import ListenBrainzClient, get_listenbrainz_client
 from ...config import Settings, get_settings
 from ...db import get_db
 from ...schemas.listens import IngestResponse, ListenIn, RecentListensResponse
