+++ sidetrack/api/api/v1/recs.py
@@ -12,7 +12,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sidetrack.common.models import UserSettings
 from sidetrack.services.candidates import generate_candidates
 from sidetrack.services.lastfm import LastfmService
-from sidetrack.services.listenbrainz import ListenBrainzService
+from sidetrack.services.listenbrainz import ListenBrainzClient
 from sidetrack.services.mb_map import recording_by_isrc
 from sidetrack.services.ranker import profile_from_spotify, rank
 from sidetrack.services.spotify import SpotifyService
@@ -44,7 +44,7 @@ async def list_recs(
     spotify_service: SpotifyService | None = None
     lastfm_service: LastfmService | None = None
     lastfm_user: str | None = None
-    lb_service: ListenBrainzService | None = None
+    lb_client: ListenBrainzClient | None = None
     lb_user: str | None = None
     if settings.spotify_recs_enabled and row.spotify_access_token:
         spotify_service = SpotifyService(client, access_token=row.spotify_access_token)
@@ -52,14 +52,14 @@ async def list_recs(
         lastfm_service = LastfmService(client, settings.lastfm_api_key)
         lastfm_user = row.lastfm_user
     elif settings.lb_cf_enabled and row.listenbrainz_user:
-        lb_service = ListenBrainzService(client)
+        lb_client = ListenBrainzClient(client)
         lb_user = row.listenbrainz_user
 
     candidates = await generate_candidates(
         spotify=spotify_service,
         lastfm=lastfm_service,
         lastfm_user=lastfm_user,
-        listenbrainz=lb_service,
+        listenbrainz=lb_client,
         listenbrainz_user=lb_user,
     )
 
@@ -116,7 +116,7 @@ async def ranked_recs(
     spotify_service: SpotifyService | None = None
     lastfm_service: LastfmService | None = None
     lastfm_user: str | None = None
-    lb_service: ListenBrainzService | None = None
+    lb_client: ListenBrainzClient | None = None
     lb_user: str | None = None
     if settings.spotify_recs_enabled and row.spotify_access_token:
         spotify_service = SpotifyService(client, access_token=row.spotify_access_token)
@@ -124,14 +124,14 @@ async def ranked_recs(
         lastfm_service = LastfmService(client, settings.lastfm_api_key)
         lastfm_user = row.lastfm_user
     elif settings.lb_cf_enabled and row.listenbrainz_user:
-        lb_service = ListenBrainzService(client)
+        lb_client = ListenBrainzClient(client)
         lb_user = row.listenbrainz_user
 
     candidates = await generate_candidates(
         spotify=spotify_service,
         lastfm=lastfm_service,
         lastfm_user=lastfm_user,
-        listenbrainz=lb_service,
+        listenbrainz=lb_client,
         listenbrainz_user=lb_user,
     )
