+++ sidetrack/services/candidates.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from .lastfm import LastfmService
-from .listenbrainz import ListenBrainzService
+from .listenbrainz import ListenBrainzClient
 from .spotify import SpotifyService
 
 
@@ -105,7 +105,7 @@ async def _lastfm_candidates(lfm: LastfmService, user: str) -> list[dict[str, An
     return out
 
 
-async def _listenbrainz_candidates(lb: ListenBrainzService, user: str) -> list[dict[str, Any]]:
+async def _listenbrainz_candidates(lb: ListenBrainzClient, user: str) -> list[dict[str, Any]]:
     """Generate candidate tracks for a ListenBrainz user."""
 
     out: list[dict[str, Any]] = []
@@ -131,7 +131,7 @@ async def generate_candidates(
     spotify: SpotifyService | None = None,
     lastfm: LastfmService | None = None,
     lastfm_user: str | None = None,
-    listenbrainz: ListenBrainzService | None = None,
+    listenbrainz: ListenBrainzClient | None = None,
     listenbrainz_user: str | None = None,
 ) -> list[dict[str, Any]]:
     """Return recommendation candidates for the given user."""
